+++ test/functional/rpc_blockchain.py
@@ -125,7 +125,6 @@ class BlockchainTest(BitcoinTestFramework):
             'mediantime',
             'pruned',
             'size_on_disk',
-            'softforks',
             'time',
             'verificationprogress',
             'warnings',
@@ -142,29 +141,6 @@ class BlockchainTest(BitcoinTestFramework):
 
         # size_on_disk should be > 0
         assert_greater_than(res['size_on_disk'], 0)
-
-        assert_equal(res['softforks'], {
-            'testdummy': {
-                'type': 'bip9',
-                'bip9': {
-                    'status': 'started',
-                    'bit': 28,
-                    'start_time': 0,
-                    # testdummy does not have a timeout so is set to the max
-                    # int64 value
-                    'timeout': 0x7fffffffffffffff,
-                    'since': 144,
-                    'statistics': {
-                        'period': 144,
-                        'threshold': 108,
-                        'elapsed': HEIGHT - 143,
-                        'count': HEIGHT - 143,
-                        'possible': True,
-                    },
-                },
-                'active': False,
-            },
-        })
 
         # pruneheight should be greater or equal to 0
         assert_greater_than_or_equal(res['pruneheight'], 0)