+++ doc/release-notes.md
@@ -4,4 +4,7 @@ Bitcoin ABC version 0.27.0 is now available from:
 
   <https://download.bitcoinabc.org/0.27.0/>
 
-This is a maintenance release with no user-visible change.
+  - The `softforks` field from the `getblockchaininfo` RPC is deprecated.
+    To keep using this field, use the `-deprecatedrpc=softforks` option.
+    Note that this field has been empty for a long time and will remain
+    empty until its eventual removal.